pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=24.1.0",
    "asyncio>=4.0.0",
    "better-profanity>=0.7.0",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "google-generativeai>=0.8.5",
    "httpx>=0.28.1",
    "opencv-python>=4.12.0.88",
    "pillow>=11.3.0",
    "profanity-check>=1.0.3",
    "psutil>=7.0.0",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.1.1",
    "python-magic>=0.4.27",
    "python-telegram-bot>=22.3",
    "requests>=2.32.5",
    "telegram>=0.0.1",
]
